fix: widen TT age from u8 to u16 to prevent wrap after 256 searches
The global age counter (AtomicU8) wrapped 255→0 after 256 go commands,
causing entries from one wrap-cycle ago to get age_distance=0 and the
same eviction priority as fresh entries. Gradual TT pollution in sessions
longer than ~3-7 games.

The ctrl word had 8 unused bits at bits 56-63. Widening age from 8 to
16 bits fills this gap (now bits 48-63). Wrap now happens after 65536
searches — effectively never.

+++ src/tt/entry.rs
@@ -24,7 +24,7 @@ pub struct TtEntry {
     pub best: Move,
     pub depth: i8,
     pub bound: Bound,
-    pub age: u8,
+    pub age: u16,
     /// Raw (uncorrected) static evaluation. Zero means "not stored".
     /// When non-zero, the search can reuse this instead of calling evaluate().
     pub raw_eval: i16,
